Backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
branca==0.3.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
folium==0.10.1
idna==2.10
inflection==0.5.1
isort==5.5.2
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.17.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pandas==0.25.3
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.3.1
seaborn==0.9.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
yapf==0.30.0
